Move Home component to pages directory
